refactor : vacancy DTO 패키지 분리 및 record 방식으로 변환
