--- src/main/java/com/catchtable/vacancy/dto/VacancyRegisterResponse.java
@@ -1,10 +1,0 @@
-package com.catchtable.vacancy.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class VacancyRegisterResponse {
-    private Long vacancyId;
-}
